+++ test/test_utils.hpp
@@ -4,6 +4,8 @@
 #include <string>
 #include <vector>
 #include <fstream>
+#include <sstream>
+#include <iostream>
 
 namespace gitter::test {